ESR's gpsd patch #4:  This is the one that nukes globals.
There is exactly one (1) global variable left after this patch.
It is called "session", and is a per-GPS-session object that looks like this:

struct session_t
{
    int debug;
    struct longlat initpos;
    int device_type;
    struct OUTDATA gNMEAdata;
};

The next step is for the device-type field to stop being an enum and
become an object pointer -- or as close to an object pointer as you get
in C, anyway.  It will refer to a driver object.