Our formula for computing estimated track error was computationally unstable. At low velocities it often yielded NaN on a 64-bit machine but a number close to 180 degress on a 32-bit machine.  Address this by (a) punting to NaN when points very close together cause rounding-error blowup in the earth_distance() formula, and (b) punting to NaN when distance between fixes is less than the horizontal fix error estimate.
The effect is that estimated track error will be NaN more often, but
otherwise much more reliable and less prone to diverge between systems
with different word sizes and/or FPU hardware.

This change cuts the observed number of diff lines generated by 'make regress'
on a 32-bit machine roughly in half.