Drop staging buffers for fix information from channel level to device level.
We had per-channel staging buffers for fix information because we used to
have a per-channel smoothing-policy bit, set by the old J command.  That
became obsolete when we figured out how to automate end-of-cycle detection.
There's no longer any point in the extra buffer copies and the per-channel
buffer allocation, so we go to using per-device buffers instead.

All regression tests pass.