test: Fix GNU makeism concerning exporting variables in Makefile.
The PYTHON env variable was being exported to child processes using
a GNU make extension.  Primarily this is used in regress-driver
to run the testers with the right python environment by querying
the python interpreter.

This patch makes explicit the carrying around of the PYTHON
environment variable down to regress-driver through the convolutions
of the Makefile.  This fixes the build using FreeBSD's make, which
doesn't support the variable export extension that GNU make does.

Signed-off-by: Eric S. Raymond <esr@thyrsus.com>