Enable the daemon lower layer to pass up a distinct return for 9-byte reads.
An arctual read error now returns ERRORI_IS, the same code that's used
for the packet sniffer failing sync.  This is a hard stop that says
gpsd should deactivate the device.

NODATA_IS is now reserved for signaling 0-length reads from the device.
This is a step towards handling those not as hard stops but as an indication
to back off long enough for serial buffers to refill.

All regression tests pass.