Do not wait for mountpoints in hotplug wrapper.
Udev does not fork, so waiting for mountpoints to appear at boot time
will block the boot process until we run into the timeout. A proper
implementation would be to fork a process to wait in the background
or to handle the not-yet-added devices in the gpsd init script.
So this is more a quick fix then the proper solution.

See http://bugs.debian.org/591324 for details.