Make the hotplug wrapper work when USBAUTO is not set.
The wrapper doesn't seem to work when USBAUTO is not set, this is
because [ -n ] gives different result than [ -n "" ], addressed by
the attached patch.

It includes also the options remapping to be consistent with the rpm
init script and sysconfig file, and adds /usr/sbin to PATH which is
needed on some systems to find gpsd in the hotplug script.

Signed-off-by: Eric S. Raymond <esr@thyrsus.com>