Bug #17638: control socket & command can return OK even if write fails
First patch to fix the ! Command

On of the symptoms of the bug 017637 are that writes to the serial port
fail. In particular

echo "&/dev/ttySAC1=01" | socat - UNIX-CONNECT:gpsd.sock

causes gpsd to do

write(7, "\1", 1) = -1 EBADF (Bad file descriptor)
write(8, "OK\n", 3)

which is wrong in two different ways:

1) It should not attempt to write to a read-only device
2) it should not print "OK" when the write fails

The attached patch tries to fix both issues for the ! command. If you
agree, I'll write a second patch that does the same treatment for the &
command.

Signed-off-by: Jon Schlueter <jon.schlueter@gmail.com>