Find {,sys/}endian.h via autoconf.
A recent commit introduced an unconditional include of endian.h, which
is not specified by POSIX and thus nonportable.  Apparently lacking a
portable way to get the needed le16toh etc. functions for dealing with
remote data from a le machine on the host, use autoconf to search for
endian.h (on Linux) and sys/endian.h (on NetBSD).  Add an explicit
failure for lacking either header.