Use a better path for chrony.sock
When gpsd is running as root expect the chrony.sock file in /var/run.
When gpsd is not, put it in /tmp/

Only root can use /var/run, but some folks do not like to run gpsd
or chronyd as root, so allow them to use /tmp.  Conforms to the LFSSTD.