Add device activation/deactivation hook
This adds possibility for a system integrator to define custom hooks that
are executed on device activation and deactivation.

Since there's no standard way to control power consumption of different gps
devices out there, the designers of embedded appliances usually provide a
system-specific way to do power management of a GPS module. This might be
e.g. a power_on sysfs node. This patch adds an opportunity for a system
integrator to transparently enable power for device as soon as any client
starts requesting data from it and to disable power on deactivation (which
normally happens 60 seconds after the last client who needed it
disconnected).

Signed-off-by: Eric S. Raymond <esr@thyrsus.com>