Reduce the amount of padding time in the regression tests.
There may have been a real (minor) problem that needed to be solved this way
at one time, but it was masking a major subtle bug in gpsd's packet handling
that we have since fixed.  We could probably set both pads to 0 on Linux
but because I remember a trouble report from NetBSD I'm being a bit
conservative.