Refactoring step.  All regression tests pass.
What's going in with this commit, and the last one, is that I'm
disentangling the logic for JSON reporting of device state from the
daemon dispatcher layer.  This means that other programs that use
libgpsd *without* necessarily going through the dispatcher layer will
be able to do so.

No logic changes.