Abandon the attempt to de-Unicodify JSON as it comes in.
This turned out to cause mike_t's bug with doubling of backslashes - praobably
doue to a subtle bug in the Python encode feature. And it seems not to be
necessary, as Python doesn't notice the difference between plain and Unicode
strings unless they contain Unicode.

Has the nice effect of simplifying the Python client code.  The only place
decoding to ASCII is now done is where a C extension needs it.