Don't assume python scripts in source tree are directly runnable.
Use ${PYTHON} to invoke scripts, because the sources have
"/usr/bin/env python", and a sane approach to the python version
incompatibility nightmare is to a) decline to have a program named
python and to b) force all python programs to pick a version at
configure time and stick with it.  (Otherwise you break installed
programs when the default changes, or have to go through some
difficult-to-prove-correct update process.)

With this, regression tests at least begin to run sanely instead of
all failing miserably (on NetBSD/pkgsrc).