In the client API, take the data buffer out of the gpsdata_t structure.
tz has shown me that we had a point of brittleness where the size
of the client's data buffer was exposed in gps_data_t. Solution:
move it (and the data-waiting count) to the client's private-data
member.  Requires a new but trivial client-library entry point to get
the buffer pointer when a client needs it.

Had to be done before API Version 5 froze, or we'd have been stuck breaking
a shipped API again.

No implications for the daemon or other tools, as this buffer is never
used there.  (Except it gets rid of several K of never-used static
buffer space in the daemon, which is a good thing.)

All regression tests pass. Code splints clean.  cgps (the only place
the API fix matters in our code) live-tests correctly.