Cut down on gpsd_hexdump_wrapper() calls.
This awkward construct seems to have triggered some sort of toolchain
bug.  The symptom is that gpsd_hexdump_level has different values depending
on whether we test it inside or outside of hex.c.

As a first step towards eliminating this global, we throw out a bunch
of hex logging that is duplicative or not really needed for production
drivers. Experimental drivers can call the hex dumper unconditionally.

The goal is to reduce the number of uses of this variable to eliminate it
withoud doing violence to the internal APIs. This is a first step.

All regression tests pass.